+++ Sources/GanitEditorUI/SheetCommands.swift
@@ -8,6 +8,7 @@ import AppKit
   func copyResult(_ sender: Any?)
   func copyFullPrecision(_ sender: Any?)
   func showInterpretation(_ sender: Any?)
+  func askAssistant(_ sender: Any?)
   func insertReference(_ sender: Any?)
   func nextProblem(_ sender: Any?)
   func previousProblem(_ sender: Any?)